additional restructure and configuration of swagger
rename dataextraction.feature package to dataextraction.elements
adaption of the swagger dockets according to the new package structure
